+++ src/screen_win_loose.hpp
@@ -35,7 +35,7 @@ public:
     );
 
     bool onAction( Action );
-    void setScore( uint16_t );
+    void setScore( uint32_t );
     virtual void render( RenderContext ) const override;
     virtual void update( const UpdateContext& ) override;
     virtual bool onMouseEvent( const MouseEvent& ) override;
